Extracted tests that really need neo4j in extra class
