+++ cpg-neo4j/src/test/kotlin/de/fraunhofer/aisec/cpg_vis_neo4j/ApplicationTest.kt
@@ -26,12 +26,11 @@
 package de.fraunhofer.aisec.cpg_vis_neo4j
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import de.fraunhofer.aisec.cpg.*
-import de.fraunhofer.aisec.cpg.graph.builder.*
+import de.fraunhofer.aisec.cpg.TranslationManager
+import de.fraunhofer.aisec.cpg.TranslationResult
 import de.fraunhofer.aisec.cpg.graph.declarations.FunctionDeclaration
 import de.fraunhofer.aisec.cpg.graph.functions
 import de.fraunhofer.aisec.cpg.graph.statements.expressions.CallExpression
-import de.fraunhofer.aisec.cpg.graph.types.*
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Paths
@@ -41,51 +40,24 @@ import kotlin.reflect.jvm.javaField
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
-import org.junit.jupiter.api.Tag
 import org.neo4j.ogm.annotation.Relationship
-import org.neo4j.ogm.config.ObjectMapperFactory.objectMapper
 import picocli.CommandLine
 
-@Tag("integration")
+fun createTranslationResult(): Pair<Application, TranslationResult> {
+    val topLevel = Paths.get("src").resolve("test").resolve("resources").toAbsolutePath()
+    val path = topLevel.resolve("client.cpp").toAbsolutePath()
+
+    val cmd = CommandLine(Application::class.java)
+    cmd.parseArgs(path.toString())
+    val application = cmd.getCommand<Application>()
+
+    val translationConfiguration = application.setupTranslationConfiguration()
+    val translationResult =
+        TranslationManager.builder().config(translationConfiguration).build().analyze().get()
+    return application to translationResult
+}
+
 class ApplicationTest {
-    private fun createTranslationResult(): Pair<Application, TranslationResult> {
-        val topLevel = Paths.get("src").resolve("test").resolve("resources").toAbsolutePath()
-        val path = topLevel.resolve("client.cpp").toAbsolutePath()
-
-        val cmd = CommandLine(Application::class.java)
-        cmd.parseArgs(path.toString())
-        val application = cmd.getCommand<Application>()
-
-        val translationConfiguration = application.setupTranslationConfiguration()
-        val translationResult =
-            TranslationManager.builder().config(translationConfiguration).build().analyze().get()
-        return application to translationResult
-    }
-
-    @Test
-    @Throws(InterruptedException::class)
-    fun testPush() {
-        val (application, translationResult) = createTranslationResult()
-
-        assertEquals(32, translationResult.functions.size)
-
-        application.pushToNeo4j(translationResult)
-
-        val sessionAndSessionFactoryPair = application.connect()
-
-        val session = sessionAndSessionFactoryPair.first
-        session.beginTransaction().use { transaction ->
-            val functions = session.loadAll(FunctionDeclaration::class.java)
-            assertNotNull(functions)
-
-            assertEquals(32, functions.size)
-
-            transaction.commit()
-        }
-
-        session.clear()
-        sessionAndSessionFactoryPair.second.close()
-    }
 
     @Test
     fun testSerializeCpgViaOGM() {
