+++ cpg-neo4j/src/test/kotlin/de/fraunhofer/aisec/cpg_vis_neo4j/Neo4JTest.kt
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2024, Fraunhofer AISEC. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *                    $$$$$$\  $$$$$$$\   $$$$$$\
+ *                   $$  __$$\ $$  __$$\ $$  __$$\
+ *                   $$ /  \__|$$ |  $$ |$$ /  \__|
+ *                   $$ |      $$$$$$$  |$$ |$$$$\
+ *                   $$ |      $$  ____/ $$ |\_$$ |
+ *                   $$ |  $$\ $$ |      $$ |  $$ |
+ *                   \$$$$$   |$$ |      \$$$$$   |
+ *                    \______/ \__|       \______/
+ *
+ */
+package de.fraunhofer.aisec.cpg_vis_neo4j
+
+import de.fraunhofer.aisec.cpg.graph.declarations.FunctionDeclaration
+import de.fraunhofer.aisec.cpg.graph.functions
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
+
+class Neo4JTest {
+    @Test
+    @Throws(InterruptedException::class)
+    fun testPush() {
+        val (application, translationResult) = createTranslationResult()
+
+        assertEquals(32, translationResult.functions.size)
+
+        application.pushToNeo4j(translationResult)
+
+        val sessionAndSessionFactoryPair = application.connect()
+
+        val session = sessionAndSessionFactoryPair.first
+        session.beginTransaction().use { transaction ->
+            val functions = session.loadAll(FunctionDeclaration::class.java)
+            assertNotNull(functions)
+
+            assertEquals(32, functions.size)
+
+            transaction.commit()
+        }
+
+        session.clear()
+        sessionAndSessionFactoryPair.second.close()
+    }
+}
